backend/requirements.txt
fastapi
uvicorn
replicate
python-multipart
pillow
requests
python-dotenv